+++ src/routes/3092/+layout.svelte
@@ -1,15 +1,19 @@
-<script lang="ts">
-	import { page } from '$app/stores';
+<script lang="ts">    
 	const navigation = [
-		{ label: 'Manual Entry', href: `${$page.url}/manual-entry` },
-		{ label: 'Randomized CTOs', href: `${$page.url}/randomized` }
+		{ label: 'Manual Entry', href: `manual-entry` },
+		{ label: 'Static Row Entry', href: `row-entry` },
+		{ label: 'Randomized CTOs', href: `randomized` }
 	];
 </script>
 
 <nav>
+    <h1 class="text-3xl ml-3 mt-5">3092</h1>
+
+    <div class="my-2">
 	{#each navigation as nav}
-		<a href={nav.href}>{nav.label}</a>
-	{/each}
+        <a class="mx-3" href="/3092/{nav.href}">{nav.label}</a>
+    {/each}
+    </div>
 </nav>
 
 <div>
